LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/text-size-adjust-expected.txt


FAIL text-size-adjust responsive to inherited text-size-adjust changes assert_equals: expected (string) "90%" but got (undefined) undefined
